requirements.txt
playwright
openai
python-dotenv
requests
schedule
selenium
